Update code to use new error processing functions

+++ source/30_download.sh
@@ -15,12 +15,11 @@ bb-download() {
 
     bb-log-info "Downloading $URL"
     wget -nv -O "$TARGET" -nc "$URL"
-    local CODE=$?
-    if (( $CODE != 0 ))
+    if bb-error?
     then
         bb-log-error "Unable to get $URL"
         rm "$TARGET"
-        return $CODE
+        return $BB_ERROR
     fi
     echo "$TARGET"
 }
